Add controller level unit test
