+++ src/main/java/com/example/interview/model/Book.java
@@ -1,8 +1,12 @@
 package com.example.interview.model;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
+@AllArgsConstructor
+@NoArgsConstructor
 public class Book {
     private int id;
     private String author;
